Added CleanCallable, let CleanRTSCallable extend it.
