+++ src/main/java/de/dagere/peass/ci/clean/CleanBuilder.java
@@ -56,7 +56,7 @@ public class CleanBuilder extends Builder implements SimpleBuildStep, Serializab
             run.setResult(Result.FAILURE);
             return;
          }
-         CleanRTSCallable.cleanFolder(resultsFolders);
+         new CleanRTSCallable(listener).cleanFolder(resultsFolders);
       } else {
          listener.getLogger().println("Regression Test Selection cleaning disabled");
       }
