+++ src/main/java/de/dagere/peass/ci/clean/callables/CleanRTSCallable.java
@@ -1,52 +1,22 @@
 package de.dagere.peass.ci.clean.callables;
 
-import java.io.File;
 import java.io.IOException;
 
 import org.apache.commons.io.FileUtils;
-import org.jenkinsci.remoting.RoleChecker;
 
-import de.dagere.peass.ci.process.JenkinsLogRedirector;
-import de.dagere.peass.folders.PeassFolders;
 import de.dagere.peass.folders.ResultsFolders;
-import hudson.FilePath.FileCallable;
 import hudson.model.TaskListener;
-import hudson.remoting.VirtualChannel;
 
-public class CleanRTSCallable implements FileCallable<Boolean> {
+public class CleanRTSCallable extends CleanCallable {
 
    private static final long serialVersionUID = 6370053198339266977L;
    
-   private final TaskListener listener;
-
-   public CleanRTSCallable(final TaskListener listener) {
-      this.listener = listener;
+   public CleanRTSCallable(TaskListener listener) {
+      super(listener);
    }
 
    @Override
-   public void checkRoles(final RoleChecker checker) throws SecurityException {
-
-   }
-
-   @Override
-   public Boolean invoke(final File potentialSlaveWorkspace, final VirtualChannel channel) throws IOException, InterruptedException {
-      try (final JenkinsLogRedirector redirector = new JenkinsLogRedirector(listener)) {
-         String projectName = potentialSlaveWorkspace.getName();
-         File folder = new File(potentialSlaveWorkspace.getParentFile(), projectName + PeassFolders.PEASS_FULL_POSTFIX);
-         ResultsFolders resultsFolders = new ResultsFolders(folder, projectName);
-
-         cleanFolder(resultsFolders);
-
-         return true;
-      } catch (IOException e) {
-         listener.getLogger().println("Exception thrown");
-         e.printStackTrace(listener.getLogger());
-         e.printStackTrace();
-         return false;
-      }
-   }
-   
-   public static void cleanFolder(final ResultsFolders resultsFolders) throws IOException {
+   public void cleanFolder(final ResultsFolders resultsFolders) throws IOException{
       deleteResultFiles(resultsFolders);
       deleteLogFolders(resultsFolders);
    }
